factor out ts compiler and rename esbuild to bundler
