+++ packages/plugins/experimental/plugin-script/project.json
@@ -28,7 +28,8 @@
       "options": {
         "forceExit": true
       }
-    }
+    },
+    "vitest": {}
   },
   "implicitDependencies": [
     "esbuild"
