+++ packages/plugins/experimental/plugin-script/src/bundler/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export * from './bundler';
