+++ packages/plugins/experimental/plugin-script/src/compiler/index.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2023 DXOS.org
+// Copyright 2024 DXOS.org
 //
 
 export * from './compiler';
