+++ packages/plugins/experimental/plugin-script/src/components/ScriptBlock/FrameContainer/FrameContainer.tsx
@@ -11,11 +11,11 @@ import { baseSurface, mx } from '@dxos/react-ui-theme';
 import { createProtoRpcPeer } from '@dxos/rpc';
 import { createIFramePort } from '@dxos/rpc-tunnel';
 
-import { type CompilerResult } from '../../../compiler';
+import { type BundlerResult } from '../../../bundler';
 
 export type FrameContainerProps = {
   containerUrl: string;
-  result: CompilerResult;
+  result: BundlerResult;
   debug?: boolean;
 };
